Move output discovery out of the UI layer so CI can test it
CI installs without the UI extras, so importing fibsem.ui pulls napari and fails.
Every test under tests/ui/ is skipped there for that reason. The new discovery
tests lived in that directory and errored on import rather than skipping, because
the module they imported was a widget module.

Guarding them with importorskip would have made them skip instead of fail, but
the discovery logic is the substance of this change and would then never run in
CI at all.

final_reference_images has no Qt in it -- it reads a lamella path and a task
record and returns paths. It moves to fibsem/applications/autolamella/task_outputs.py,
which the review panel imports. Its tests move alongside, into tests/autolamella/,
where CI runs them.

The coincidence-viewer test genuinely needs the widget module, so it stays under
tests/ui/ and now guards on both PyQt5 and napari.

Full suite: 997 passed, 7 skipped.
